Add cache to `isGitIgnored` and `isGitIgnoredSync`
